Crate namada_tx

Namada transaction.

Re-exports

pub use event::new_tx_event;

pub use either;

Modules

action

Actions applied in a transaction.

data

Data-Types that are used in transactions.

event

Transaction events.

proto

Generated protobuf tx type.

Structs

Authorization

A section representing a multisig over another section

BatchedTx

A transaction with the commitment to a specific inner transaction of the batch

BatchedTxRef

A reference to a transaction with the commitment to a specific inner transaction of the batch

Code

A section representing transaction code

CompressedAuthorization

A section representing a multisig over another section

Data

A section representing transaction data

Header

A Namada transaction header indicating where transaction subcomponents can be found

IndexedTx

Represents the pointers to a indexed tx, which are the block height and the index inside that block. Optionally points to a specific inner tx inside a batch if such level of granularity is required.

IndexedTxRange

Inclusive range of IndexedTx entries.

MaspBuilder

A section providing the auxiliary inputs used to construct a MASP transaction

SignableEthMessage

Tag type that indicates we should use ABI serialization to sign data in a Signable wrapper.

SignatureIndex

Signature index within a multisig

Signed

A generic signed data wrapper for serialize-able types.

Tx

A Namada transaction is represented as a header followed by a series of sections providing additional details.

TxCommitments

An inner transaction of the batch, represented by its commitments to the Code, Data and Memo sections

Enums

Commitment

Represents either some code bytes or their SHA-256 hash

DecodeError

Section

A section of a transaction. Carries an independent piece of information necessary for the processing of a transaction.

Signer

Indicates the list of public keys against which signatures will be verified

TxError

VerifySigError

Represents an error in signature verification

Constants

SALT_LENGTH

Length of the transaction sections salt

Functions

standalone_signature

Get a signature for data

verify_standalone_sig

Verify that the input data has been signed by the secret key counterpart of the given public key.

Type Aliases

Memo

A memo field (bytes).