namada_tx::data
Module pos

txs to manage pos Types used for PoS system transactions

Structs

BecomeValidator

A tx data type to become a validator account.

Bond

A bond is a validator’s self-bond or a delegation from non-validator to a validator.

ClaimRewards

A claim of pending rewards.

CommissionChange

A change to the validator commission rate.

ConsensusKeyChange

A change to the validator’s consensus key.

MetaDataChange

A change to the validator metadata.

Redelegation

A redelegation of bonded tokens from one validator to another.

Withdraw

A withdrawal of an unbond.

Type Aliases

Unbond

An unbond of a bond.