namada_tx
Function verify_standalone_sig

pub fn verify_standalone_sig<T, S: Signable<T>>(
    data: &T,
    pk: &PublicKey,
    sig: &Signature,
) -> Result<(), VerifySigError>

Verify that the input data has been signed by the secret key counterpart of the given public key.